+++ lib/home.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/Login.dart';
-import 'package:flutter_app/Configure.dart';
+import 'package:flutter_app/login.dart';
+import 'package:flutter_app/configure.dart';
 import 'package:flutter_app/control.dart';
-import 'package:flutter_app/AddUser.dart';
-import 'package:flutter_app/AddGesture.dart';
-import 'package:flutter_app/History.dart';
+import 'package:flutter_app/add_user.dart';
+import 'package:flutter_app/add_gesture.dart';
+import 'package:flutter_app/history.dart';
 
 
 class Home extends StatefulWidget {
